+++ src/locale.lua
@@ -1,4 +1,6 @@
-local _, Addon = ...
+local Addon = select(2, ...) ---@type Addon
+
+--- @class Locale
 local L = Addon:GetModule("Locale")
 
 setmetatable(L, {
@@ -12,36 +14,28 @@ setmetatable(L, {
 -- ============================================================================
 
 L["ANCHOR"] = "Anchor"
-L["COMMAND_DESCRIPTION_ALPHA_RESET"] = "Reset notification background transparency."
-L["COMMAND_DESCRIPTION_ALPHA"] = "Set notification background transparency (from %s to %s percent)."
-L["COMMAND_DESCRIPTION_ANCHOR_RESET"] = "Reset the notifications anchor."
-L["COMMAND_DESCRIPTION_ANCHOR"] = "Toggle the notifications anchor."
-L["COMMAND_DESCRIPTION_DELAY_RESET"] = "Reset the delay time before notifications fade out."
-L["COMMAND_DESCRIPTION_DELAY"] = "Set the delay time before notifications fade out (from %s to %s seconds)."
-L["COMMAND_DESCRIPTION_HELP"] = "Display a list of commands."
-L["COMMAND_DESCRIPTION_MAX_RESET"] = "Reset the maximum number of displayed notifications."
-L["COMMAND_DESCRIPTION_MAX"] = "Set the maximum number of displayed notifications (from %s to %s)."
-L["COMMAND_DESCRIPTION_MONEY"] = "Toggle money notifications."
-L["COMMAND_DESCRIPTION_PRICES"] = "Toggle loot prices in notifications."
-L["COMMAND_DESCRIPTION_SPACE_RESET"] = "Reset the amount of spacing between displayed notifications."
-L["COMMAND_DESCRIPTION_SPACE"] = "Set the amount of spacing between displayed notifications (from %s to %s)."
-L["COMMAND_DESCRIPTION_TEST"] = "Test notifications."
-L["COMMAND_EXAMPLE_USAGE"] = "Example Usage"
-L["COMMAND_SUCCESS_ALPHA_RESET"] = "Notification background transparency has been reset."
-L["COMMAND_SUCCESS_ALPHA"] = "Notification background transparency set to %s%%."
-L["COMMAND_SUCCESS_ANCHOR_RESET"] = "The notifications anchor has been reset."
-L["COMMAND_SUCCESS_DELAY_RESET"] = "The delay time before notifications fade out has been reset."
-L["COMMAND_SUCCESS_DELAY"] = "Notification fade-out delay set to %s seconds."
-L["COMMAND_SUCCESS_MAX_RESET"] = "The maximum number of displayed notifications has been reset."
-L["COMMAND_SUCCESS_MAX"] = "Maximum number of displayed notifications set to %s."
-L["COMMAND_SUCCESS_SPACE_RESET"] = "The amount of spacing between displayed notifications has been reset."
-L["COMMAND_SUCCESS_SPACE"] = "Notification spacing set to %s."
-L["LOOT_PRICES_DISABLED"] = "Loot prices disabled."
-L["LOOT_PRICES_ENABLED"] = "Loot prices enabled."
-L["MAXIMUM"] = "Maximum"
-L["MINIMUM"] = "Minimum"
-L["MONEY_NOTIFICATIONS_DISABLED"] = "Money notifications disabled."
-L["MONEY_NOTIFICATIONS_ENABLED"] = "Money notifications enabled."
+L["GENERAL"] = "General"
+L["LEFT_CLICK"] = "Left-Click"
+L["LOOT_PRICES_TEXT"] = "Loot Prices"
+L["LOOT_PRICES_TOOLTIP"] = "Include the vendor price of items in loot notifications."
+L["MAX_NOTIFICATIONS_TEXT"] = "Max Notifications"
+L["MAX_NOTIFICATIONS_TOOLTIP"] = "The maximum number of displayed notifications."
+L["MONEY_NOTIFICATIONS_TEXT"] = "Money Notifications"
+L["MONEY_NOTIFICATIONS_TOOLTIP"] = "Show a notification when money is looted or received."
+L["NOTIFICATION_ALPHA_TEXT"] = "Notification Alpha"
+L["NOTIFICATION_ALPHA_TOOLTIP"] = "The background transparency of notifications."
+L["NOTIFICATION_FADE_OUT_DELAY_TEXT"] = "Notification Fade-Out Delay"
+L["NOTIFICATION_FADE_OUT_DELAY_TOOLTIP"] = "The delay time before notifications fade out."
+L["NOTIFICATION_SPACING_TEXT"] = "Notification Spacing"
+L["NOTIFICATION_SPACING_TOOLTIP"] = "The amount of space between notifications."
+L["OPTIONS_TEXT"] = "Options"
+L["OWNED_ITEM_COUNTS_TEXT"] = "Owned Item Counts"
+L["OWNED_ITEM_COUNTS_TOOLTIP"] = "Include an owned item count in loot notifications."
+L["RESET_ANCHOR"] = "Reset Anchor"
+L["RIGHT_CLICK"] = "Right-Click"
+L["SHIFT_KEY"] = "Shift"
+L["TEST_NOTIFICATIONS"] = "Test Notifications"
+L["TOGGLE_ANCHOR"] = "Toggle Anchor"
 
 -- ============================================================================
 -- Others